% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_stl_wards.R
\docType{data}
\name{ar_stl_wards}
\alias{ar_stl_wards}
\title{Ward Boundaries in St. Louis, 2010}
\format{
A data frame with 28 rows and 4 variables:
\describe{
  \item{OBJECTID}{Artifact from ESRI data creation}
  \item{WARD}{Ward number}
  \item{AREA}{area of each ward}
  \item{geometry}{simple features geometry}
  }
}
\source{
City of St. Louis
}
\usage{
data(ar_stl_wards)
}
\description{
A simple features data set containing the 2010 Ward boundaries, which
    are used as districts for Alderpersons who serve as elected representatives.
    The \code{OBJECTID} and \code{AREA} columns are included to simulate "real"
    data that may have superfluous or unclear columns.
}
\examples{
str(ar_stl_wards)
head(ar_stl_wards)
summary(ar_stl_wards$AREA)

}
\keyword{datasets}
