% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_tessellate.R
\name{ar_tessellate}
\alias{ar_tessellate}
\title{Create Tessellations From SF Object}
\usage{
ar_tessellate(.data, shape = "square", size = 1)
}
\arguments{
\item{.data}{An object of class \code{sf} to tessellate from}

\item{shape}{One of 'square' or 'hexagon', the shape to make tessellations from}

\item{size}{Numeric multiplier for size of tessellations, default is one kilometer}
}
\value{
A \code{sf} object
}
\description{
Create Tessellations From SF Object
}
\examples{
ar_tessellate(ar_stl_wards)

ar_tessellate(ar_stl_wards, shape = "hexagon", size = .75)

}
