% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_validate.R
\name{ar_validate_crs}
\alias{ar_validate_crs}
\title{Testing for shared coordinates for source and target data}
\usage{
ar_validate_crs(source, target)
}
\arguments{
\item{source}{A \code{sf} object with data to be interpolated}

\item{target}{A \code{sf} object that data should be interpolated to}
}
\value{
A logical scalar; if \code{TRUE}, the test is passed.
}
\description{
\code{awrvalidate_crs} conducts a logic test for shared coordinate
    coordinate systems, which are a requirement for interpolation.
}
