% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_aggregate.R
\name{aw_aggregate}
\alias{aw_aggregate}
\title{Aggregate new field of interest by target IDs}
\usage{
aw_aggregate(.data, target, tid, newField)
}
\arguments{
\item{target}{A given target dataset}

\item{tid}{A given target id}

\item{newField}{A given new esimation field}

\item{intersection}{A given intersected dataset}
}
\value{
A target dataset with a new field for properly calculated field of interest by target IDs
}
\description{
\code{aw_aggregate()} Distributes new field estimate values by target IDs and is joined to target dataset
}
