% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_area_wght.R
\name{aw_area_wght}
\alias{aw_area_wght}
\title{Calculate area weight using source ID areas}
\usage{
aw_area_wght(intersection)
}
\arguments{
\item{intersection}{A given intersected dataset}
}
\value{
An intersected file of class sf with field for area weight
}
\description{
\code{aw_area_wght()} This function creates an area weight field
by dividing intersection slivers by source ID areas
}
