% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_calculate.R
\name{aw_calculate}
\alias{aw_calculate}
\title{Make field new variable}
\usage{
aw_calculate(.data, newField, vals, areaWeight)
}
\arguments{
\item{.data}{A given intersected dataset}

\item{newField}{A new field name}

\item{vals}{A given variable of estimations to perform interpolation calculations on}
}
\value{
An intersected file of class sf with a new field of interest recalculated with area weight
}
\description{
\code{aw_calculate()} Recalculates existing field by area weight and yields new field
}
