% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_fieldnew.R
\name{aw_fieldnew}
\alias{aw_fieldnew}
\title{Make field new variable}
\usage{
aw_fieldnew(intersection)
}
\arguments{
\item{intersection}{A given intersected dataset}
}
\value{
An intersected file of class sf with a new field of interest recalculated with area weight
}
\description{
\code{aw_fieldnew()} Recalculates existing field by area weight and yields new field
}
