% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_intersect.R
\name{aw_intersect}
\alias{aw_intersect}
\title{Intersect Source and Target Data}
\usage{
aw_intersect(.data, source, areaVar)
}
\arguments{
\item{source}{A given source dataset}

\item{areaVar}{A given area variable}

\item{target}{A given target dataset}
}
\value{
An intersected object of class sf with calculated geometric area field
}
\description{
\code{aw_intersect()} This function intersects source and target datasets and computes a new geometric
area field in the intersection
}
