% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_stl_asthma.R
\docType{data}
\name{aw_stl_asthma}
\alias{aw_stl_asthma}
\title{Asthma in St. Louis by Census Tract, 2017}
\format{A data frame with 106 rows and 24 variables:
\describe{
  \item{GEOID}{full GEOID string}
  \item{STATEFP}{state FIPS code}
  \item{COUNTYFP}{county FIPS code}
  \item{TRACTCE}{tract FIPS code}
  \item{NAMELSAD}{tract name}
  \item{ALAND}{area of tract land, square meters}
  \item{AWATER}{area of tract water, square meters}
  \item{ASTHMA}{percent of residents with current asthma diagnosis, estimated}
  \item{geometry}{simple features geometry}
  }}
\source{
Centers for Disease Control's 500 Cities Data
}
\usage{
data(aw_stl_asthma)
}
\description{
A simple features data set containing the geometry and asthma estimates
    from the Centers for Disease Control for St. Louis.
}
\examples{
str(aw_stl_asthma)
head(aw_stl_asthma)
summary(aw_stl_asthma$ASTHMA)

}
\keyword{datasets}
