% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_intersect.R
\name{aw_strip_units}
\alias{aw_strip_units}
\title{Strip variable of attached units}
\usage{
aw_strip_units(.data, areaVar)
}
\arguments{
\item{.data}{Dataframe that variable to strip units from is located}

\item{areaVar}{A given variable to strip units from}
}
\value{
A \code{sf} object with the new area field as a numeric class.
}
\description{
\code{aw_strip_units} strips a given variable of its
    units class and instead returns a numeric value. This is called
    by \code{aw_intersect}.
}
