% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_sum.R
\name{aw_sum}
\alias{aw_sum}
\title{Create new total area field in intersection}
\usage{
aw_sum(.data, sid, areaVar, totalVar)
}
\arguments{
\item{.data}{A given dataframe}

\item{sid}{A given source id field}

\item{areaVar}{A given area variable}

\item{totalVar}{A new total area field to be estimated}
}
\value{
An intersection file of class sf with a new field for total area by source id
}
\description{
\code{aw_sum()} Produces new total area field that reflects total area by source id
}
