% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_total_id_area.R
\name{aw_total_id_area}
\alias{aw_total_id_area}
\title{Total Area by Source IDs}
\usage{
aw_total_id_area(intersection)
}
\arguments{
\item{intersection}{A given intersected dataset}
}
\value{
An intersected file of class sf with field for ID area
}
\description{
\code{aw_total_id_area()} This function logically tests for sf object status
between source and target data. Output is either TRUE for shared sf object status or FALSE.
}
