% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate_data.R
\name{aw_validate_data}
\alias{aw_validate_data}
\title{Testing for sf object status, shared unit type, and coordinate type for source
and target data.}
\usage{
aw_validate_data(source, target)
}
\arguments{
\item{source}{A given source dataset}

\item{target}{A given target dataset}
}
\value{
A logical value
}
\description{
\code{aw_validate_data()} This function logically tests for sf object status,
shared unit types, and shared coordinates between source and target data. Output is either TRUE
if all the test results are TRUE or FALSE if any individual test is FALSE.
}
