% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate.R
\name{aw_validate_longlat}
\alias{aw_validate_longlat}
\title{Testing for type of coordinates}
\usage{
aw_validate_longlat(.data)
}
\arguments{
\item{.data}{A sf object}
}
\value{
A logical scalar; if \code{TRUE}, the test is passed
}
\description{
\code{aw_validate_longlat} conducts a logic test for
    whether or not the data are in planar format.
}
