% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate.R
\name{aw_validate_sf}
\alias{aw_validate_sf}
\title{Testing for sf object status for source and target data}
\usage{
aw_validate_sf(source, target)
}
\arguments{
\item{source}{A given source dataset}

\item{target}{A given target dataset}
}
\value{
A logical value output
}
\description{
\code{aw_validate_sf()} This function logically tests for shared
sf status. Output is either TRUE for shared unit type or FALSE if sf class is not present
in at least one file.
}
