% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate.R
\name{aw_validate_units}
\alias{aw_validate_units}
\title{Testing for shared unit type status for source and target data}
\usage{
aw_validate_units(source, target)
}
\arguments{
\item{source}{A \code{sf} object with data to be interpolated}

\item{target}{A \code{sf} object that data should be interpolated to}
}
\value{
A logical scalar; if \code{TRUE}, the test is passed.
}
\description{
\code{aw_validate_units} conducts a logic test for shared
    measurement units, which are a requirement for interpolation. This
    should not be an issue if \link{aw_validate_crs} is passed, and
    may not be included in the release version.
}
