% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate.R
\name{aw_validate_vars}
\alias{aw_validate_vars}
\title{Testing for Variable Conflicts}
\usage{
aw_validate_vars(.data, varList)
}
\arguments{
\item{.data}{A sf object}

\item{varList}{A vector of variables to be created}
}
\value{
A logical scalar; if \code{TRUE}, the test is passed
}
\description{
\code{aw_validate_vars} conducts a logic test for
    whether or not any of the variables to be created in the target
    data already exist as named columns.
}
