% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_weight.R
\name{aw_weight}
\alias{aw_weight}
\title{Calculate area weight using source ID areas}
\usage{
aw_weight(.data, areaVar, totalVar, areaWeight)
}
\arguments{
\item{.data}{A given intersected dataset}

\item{areaVar}{A given area variable}

\item{totalVar}{A given total area field estimated by source id}

\item{areaWeight}{A given name for the area weight calculation field}
}
\value{
An intersected file of class sf with field for area weight
}
\description{
\code{aw_weight()} This function creates an area weight field
by dividing intersection area field by total area field
}
