% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_aggregate.R
\name{aw_aggregate}
\alias{aw_aggregate}
\title{Aggregate Estimates Based on Target ID}
\usage{
aw_aggregate(.data, target, tid, interVar, newVar)
}
\arguments{
\item{.data}{A given intersected dataset}

\item{target}{A \code{sf} object that data should be interpolated to}

\item{tid}{A unique identification number within \code{target}}

\item{interVar}{A variable containing an interpolated value created by \code{aw_calculate}}

\item{newVar}{Optional; a new field name to store the interpolated value in. If not specified,
the \code{interVar} argument will be used as the new field name.}
}
\value{
A \code{sf} object with the interpolated value added to it.
}
\description{
\code{aw_aggregate} sums the new estimates produced by \link{aw_calculate}
    based on the target id. These are then joined with the target data. This is
    the fourth step in the interpolation process after \link{aw_weight}.
}
