% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_interpolate.R
\name{aw_interpolate}
\alias{aw_interpolate}
\title{Interpolate Values}
\usage{
aw_interpolate(.data, tid, source, sid, type = "extensive", output = "sf", ...)
}
\arguments{
\item{.data}{A \code{sf} object that data should be interpolated to (this is referred
to as the \code{target} elsewhere in the package).}

\item{tid}{A unique identification number within \code{target}}

\item{source}{A \code{sf} object with data to be interpolated}

\item{sid}{A unique identification number within \code{source}}

\item{type}{One of either \code{"extensive"} (if the data are spatitally extensive e.g.
population counts) or \code{"intensive"} (if the data are spatially intensive e.g.
population density). If \code{"extensive"}, the sum is returned for the interpolated
value. If \code{"intensive"}, the mean is returned for the interpolated value.}

\item{output}{If \code{"tibble"}, will return a tibble instead of an \code{sf} object.}

\item{...}{A list of columns from \code{source}, with each name quoted, that should
interpolated into the \code{target} data (these are referred to as the \code{value}
elsewhere in the package).}
}
\value{
A \code{sf} object or a \code{tibble} with the value or values interpolated into
    the \code{target} data.
}
\description{
This is the core function within the package. It evalues both the input
    data sources and validates them before interpolating one or more listed values
    from the source data into the target data.
}
\details{
Areal weighted interpolation can be used for generating demographic
    estimates for overlapping but incongruent polygon features. It assumes that
    individual members of a population are evenly dispersed within the source features
    (an assumption not likely to hold in the real world). It also functions best
    when data are in a projected coordinate system, like the UTM coordinate system.
}
