% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_interpolate.R
\name{aw_interpolate}
\alias{aw_interpolate}
\title{Interpolate}
\usage{
aw_interpolate(.data, tid, source, sid, output = "sf", ...)
}
\arguments{
\item{.data}{A \code{sf} object that data should be interpolated to (this is referred
to as the \code{target} elsewhere in the package).}

\item{tid}{A unique identification number within \code{target}}

\item{source}{A \code{sf} object with data to be interpolated}

\item{sid}{A unique identification number within \code{source}}

\item{output}{If \code{"tibble"}, will return a tibble instead of an \code{sf} object.}

\item{...}{A list of columns from \code{source}, with each name quoted, that should
interpolated into the \code{target} data (these are referred to as the \code{value}
elsewhere in the package).}
}
\value{
A \code{sf} object or a \code{tibble} with the value or values interpolated into
    the \code{target} data.
}
\description{
Interpolate
}
