% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_intersect.R
\name{aw_strip_units}
\alias{aw_strip_units}
\title{Strip dataframe variable of attached units}
\usage{
aw_strip_units(.data, areaVar)
}
\arguments{
\item{.data}{Dataframe that variable to strip units from is located}

\item{var}{A given variable to strip units from}
}
\value{
A dataframe with a variable stripped of attached units
}
\description{
\code{aw_strip_units()} Strips dataframe variable of unit class type
}
