% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_sum.R
\name{aw_sum}
\alias{aw_sum}
\title{Create new total area field in intersection}
\usage{
aw_sum(.data, sid, areaVar, totalVar)
}
\arguments{
\item{.data}{A \code{sf} object that data should be interpolated to}

\item{sid}{A unique identification number within \code{source}}

\item{areaVar}{The name of the a variable measuring a feature's area}

\item{totalVar}{The name of a new total area field to be calculated}
}
\value{
A \code{sf} object with the intersected data and new total area field.
}
\description{
\code{aw_sum} produces a new total area field that contains
    the total area by \code{source} id. This is the second step in the
    interpolation process after \link{aw_intersect}.
}
