% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate.R
\name{aw_validate}
\alias{aw_validate}
\title{Valdating Data for Interpolation}
\usage{
aw_validate(source, target, verbose = FALSE)
}
\arguments{
\item{source}{A \code{sf} object with data to be interpolated}

\item{target}{A \code{sf} object that data should be interpolated to}

\item{verbose}{A logical scalar; if \code{TRUE}, a tibble with test results is returned}
}
\value{
If \code{verbose} is \code{FALSE}, a logical scalar is returned that is \code{TRUE}
    is all tests are passed and \code{FALSE} if one or more tests is failed. If \code{verbose}
    is \code{TRUE}, a tibble with detailed test results is returned.
}
\description{
\code{aw_validate} executes a series of logic tests for \code{sf} object status,
shared unit types, and shared coordinates between source and target data.
}
