% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate_crs.R
\name{aw_validate_crs}
\alias{aw_validate_crs}
\title{Testing for shared coordinates for source and target data}
\usage{
aw_validate_crs(source, target)
}
\arguments{
\item{source}{A given source dataset}

\item{target}{A given target dataset}
}
\value{
A logical value
}
\description{
\code{aw_validate_crs()} This function logically tests for shared coordinates
between source and target data. Output is either TRUE for shared coordinates or FALSE if they differ.
}
