% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_validate.R
\name{aw_validate_units}
\alias{aw_validate_units}
\title{Testing for shared unit type status for source and target data}
\usage{
aw_validate_units(source, target)
}
\arguments{
\item{source}{A given source dataset}

\item{target}{A given target dataset}
}
\value{
A logical value output
}
\description{
\code{aw_validate_units()} This function logically tests for shared
unit type status. Output is either TRUE for shared unit type or FALSE if unit types
differ.
}
